+++ amp/analysis.py
@@ -9,6 +9,8 @@ matplotlib.use('Agg')
 from matplotlib import pyplot
 from matplotlib.backends.backend_pdf import PdfPages
 from matplotlib import rcParams
+from ase.io import Trajectory
+
 rcParams.update({'figure.autolayout': True})
 
 
@@ -260,8 +262,27 @@ def plot_parity_and_error(calc,
     energy_data = {}
     for hash, image in images.iteritems():
         no_of_atoms = len(image)
-        amp_energy = calc.model.calculate_energy(
-            calc.descriptor.fingerprints[hash])
+        energy_args = dict(
+                fingerprints=calc.descriptor.fingerprints[hash],
+                )
+
+        model_name = calc.model.__class__.__name__
+
+        if model_name == 'KernelRidge':
+            if calc.model.trainingimages is not None:
+                trainingimages = hash_images(Trajectory(calc.model.trainingimages))
+                energy_args['trainingimages'] = trainingimages
+                calc.descriptor.calculate_fingerprints(
+                        images=trainingimages,
+                        parallel=calc._parallel,
+                        log=calc._log,
+                        calculate_derivatives=calculate_derivatives
+                        )
+                fp_trainingimages = calc.descriptor.fingerprints
+                energy_args['fp_trainingimages'] = fp_trainingimages
+                energy_args['hash'] = hash
+
+        amp_energy = calc.model.calculate_energy(**energy_args)
         actual_energy = image.get_potential_energy(apply_constraint=False)
         act_energy_per_atom = actual_energy / no_of_atoms
         energy_error = abs(amp_energy - actual_energy) / no_of_atoms
@@ -289,10 +310,24 @@ def plot_parity_and_error(calc,
         calc._log('Calculating forces...', tic='forces')
         force_data = {}
         for hash, image in images.iteritems():
-            amp_forces = \
-                calc.model.calculate_forces(
-                    calc.descriptor.fingerprints[hash],
-                    calc.descriptor.fingerprintprimes[hash])
+            forces_args = dict(
+                    fingerprints=calc.descriptor.fingerprints[hash],
+                    fingerprintprimes=calc.descriptor.fingerprintprimes[hash]
+                    )
+
+            if model_name == 'KernelRidge':
+                if calc.model.trainingimages is not None:
+                    trainingimages = \
+                            hash_images(Trajectory(calc.model.trainingimages))
+                    calc.descriptor.calculate_fingerprints(
+                            images=trainingimages,
+                            calculate_derivatives=True
+                            )
+                    t_descriptor = calc.descriptor
+                    forces_args['trainingimages'] = trainingimages
+                    forces_args['t_descriptor'] = t_descriptor
+
+            amp_forces =  calc.model.calculate_forces(**forces_args)
             actual_forces = image.get_forces(apply_constraint=False)
             force_data[hash] = [actual_forces, amp_forces,
                                 abs(np.array(amp_forces) -