+++ amp/regression/__init__.py
@@ -2,7 +2,6 @@ from ..utilities import ConvergenceOccurred
 
 
 class Regressor:
-
     """Class to manage the regression of a generic model. That is, for a
     given parameter set, calculates the cost function (the difference in
     predicted energies and actual energies across training images), then
@@ -28,7 +27,7 @@ class Regressor:
         """optimizer can be specified; it should behave like a
         scipy.optimize optimizer. That is, it should take as its first two
         arguments the function to be optimized and the initial guess of the
-        optimal paramters. Additional keyword arguments can be fed through
+        optimal parameters. Additional keyword arguments can be fed through
         the optimizer_kwargs dictionary."""
 
         user_kwargs = optimizer_kwargs
@@ -55,7 +54,12 @@ class Regressor:
         elif optimizer == 'NCG':
             from scipy.optimize import fmin_ncg as optimizer
             optimizer_kwargs = {'avextol': 1e-15, }
-
+        elif optimizer == 'fmin':
+            from scipy.optimize import fmin as optimizer
+            optimizer_kwargs = {
+                    'maxfun': 99999999,
+                    'maxiter': 9999999999
+                    }
         if user_kwargs:
             optimizer_kwargs.update(user_kwargs)
         self.optimizer = optimizer
@@ -85,7 +89,6 @@ class Regressor:
                                **self.optimizer_kwargs)
             else:
                 self.optimizer(model.get_loss, x0, **self.optimizer_kwargs)
-
         except ConvergenceOccurred:
             log('...optimization successful.', toc='opt')
             return True